refactor(settings): separate AppSettings and move SettingsService
